+++ node/src/Transaction.ts
@@ -107,6 +107,7 @@ import {
     createXLen,
     createZInterCard,
     createObjectIdletime,
+    createObjectRefcount,
 } from "./Commands";
 import { redis_request } from "./ProtobufMessage";
 
@@ -1506,6 +1507,20 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public objectIdletime(key: string): T {
         return this.addAndReturn(createObjectIdletime(key));
+    }
+
+    /**
+     * Returns the reference count of the object stored at `key`.
+     *
+     * See https://valkey.io/commands/object-refcount/ for more details.
+     *
+     * @param key - The `key` of the object to get the reference count of.
+     *
+     * Command Response - If `key` exists, returns the reference count of the object stored at `key` as a `number`.
+     * Otherwise, returns `null`.
+     */
+    public objectRefcount(key: string): T {
+        return this.addAndReturn(createObjectRefcount(key));
     }
 }